LayoutTests/fast/dom/StyleSheet/gc-parent-rule.html

<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body onload="test()">
<script>
window.jsTestIsAsync = true;
description("Test that CSS rules don't lose custom properties on GC when only reachable through child rules.");

function test()
{
    var sheet = document.implementation.createCSSStyleSheet("title", "screen");
    sheet.insertRule("@media all { div { color:black } }", 0);
    rule = sheet.rules[0].cssRules[0];
    rule.parentRule.foo = "bar";
    sheet = null;

    gc();
    setTimeout(function() {
        gc();
        shouldBe('rule.parentRule.foo', '"bar"');
        finishJSTest();
    }, 0);
}
</script>
<foo id="mystyle" style="display:none">.div { color: white; }</foo>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
